New approach to the import of RCS commits.
The commit graph for an RCS file is now generated first,
and then the tags are attached.
COMMIT_FAKE is no more; it has been replaced with the RCS state "fake".
Raking of dead leaves is now done on import and not in a separate pass.
This solves several issues with invalid sets of dead leaves.