Roxen: Fixed excessive renaming.
The bug was due to not considering that append_revision()
returns 0 (zero) for commit times before the root commit.
This zero in turn indicates to rename_revision()
that it should rename the entire set of revisions,
when actually in this case the opposite is desired.

This bug caused several files to appear in obsolete directories,
when in fact they had never been there.